\docType{data}
\name{stocks}
\alias{stocks}
\title{Daily Closing Prices of Major European Stock Indices, 1991 to 1998}
\format{A data.frame with 7440 rows and 3 variables}

\description{
  A dataset containing the daily closing prices of four European Stock Indices between 1991 and 1998. Adapted from EuStockMarkets.
}
\usage{
  stocks
}
\details{
  \itemize{ \item x. Timestamp in consecutive order.
    \item group. Name of the stock.
    \item y. Closing price on that day.
  }
}
\source{
  The data were kindly provided by Erste Bank AG, Vienna, Austria.
}

\keyword{datasets}

